+++ src/pow.h
@@ -17,8 +17,6 @@ unsigned int GetNextWorkRequired(const CBlockIndex* pindexLast, const CBlockHead
 /** Check whether a block hash satisfies the proof-of-work requirement specified by nBits */
 bool CheckProofOfWork(uint256 hash, unsigned int nBits);
 
-void UpdateTime(CBlockHeader* block, const CBlockIndex* pindexPrev);
-
 uint256 GetProofIncrement(unsigned int nBits);
 
 #endif // BITCOIN_POW_H
